feat: trainer completing training updates currency expiries for trainees who completed the training feat: trainee withdrawal during warning period records the withdrawal feat: all updates in traineesController sets updatedAt to the current date

+++ server/src/controllers/traineesController.ts
@@ -1,4 +1,4 @@
-import { Account } from "../constants";
+import { Account, MONTHS_TO_RECORD_WITHDRAWAL } from "../constants";
 import { Request, Response } from "express";
 import { prisma } from "../config/database";
 import dayjs from "dayjs";
@@ -159,9 +159,27 @@ const updateBooking = async (req: Request, res: Response) => {
 const book = async (traineeId: number, trainingId: number) => {
   const existingBooking = await prisma.traineeToTraining.findFirst({
     where: { trainee: traineeId, training: trainingId },
+    include: {
+      trainings: {
+        select: {
+          start: true,
+        },
+      },
+    },
   });
   if (existingBooking) {
     console.log("delete booking");
+    if (
+      dayjs()
+        .add(MONTHS_TO_RECORD_WITHDRAWAL, "months")
+        .isAfter(dayjs(existingBooking.trainings.start))
+    ) {
+      console.log("record withdrawal");
+      return await prisma.traineeToTraining.update({
+        where: { id: existingBooking.id },
+        data: { status: 4, updatedAt: dayjs().toDate() },
+      });
+    }
     return await prisma.traineeToTraining.delete({
       where: { id: existingBooking.id },
     });
@@ -196,8 +214,8 @@ const update = async (req: Request, res: Response) => {
       where: { id: c.id || 0 },
       update: {
         expiry: c.expiry,
-        updatedAt: dayjs().toDate(),
         seniority: c.seniority,
+        updatedAt: dayjs().toDate(),
       },
       create: {
         expiry: c.expiry,
@@ -219,7 +237,7 @@ const update = async (req: Request, res: Response) => {
 
   const updateTrainee = prisma.trainee.update({
     where: { id: Number(traineeId) },
-    data: { category: trainee.category },
+    data: { category: trainee.category, updatedAt: dayjs().toDate() },
   });
 
   try {
@@ -296,7 +314,7 @@ const checkin = async (req: Request, res: Response) => {
             training: Number(trainingId),
           },
         },
-        data: { status: 2 },
+        data: { status: 2, updatedAt: dayjs().toDate() },
       });
       return res.status(200).json({ message: "Check in successful!" });
     } else {
